[Fix] DI 모듈 수정
모듈 컴포넌트 변경
변경된 Data, Domain에 따라 Provide 수정
