+++ domain/src/main/java/com/ddd4/dropit/domain/usecase/GetItemUseCase.kt
@@ -1,15 +1,14 @@
 package com.ddd4.dropit.domain.usecase
 
-import com.ddd4.dropit.domain.repository.DropitRepository
-import com.ddd4.dropit.domain.Result
 import com.ddd4.dropit.domain.entity.DomainEntity
+import com.ddd4.dropit.domain.repository.DropitRepository
 import java.lang.Exception
 
-class GetItemUseCase(
+class getItemUseCase(
     private val dropitRepository: DropitRepository
 ) {
 
-    suspend fun execute(folderId: Long): Result<List<DomainEntity.Item>> =
+    suspend fun execute(folderId: Long): Result =
         try {
             dropitRepository.getItems(folderId).let { data ->
                 Result.Success(data)
@@ -17,4 +16,9 @@ class GetItemUseCase(
         } catch (e: Exception) {
             Result.Error(e)
         }
+
+    sealed class Result {
+        data class Success(val data: List<DomainEntity.Item>) : Result()
+        data class Error(val e: Throwable) : Result()
+    }
 }
